refactor: 테스트에 필요한 객체 용도를 명확하게 알 수 있도록 이동
