+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMappingFeatureTest.java
@@ -17,7 +17,7 @@ class AnnotationHandlerMappingFeatureTest {
 
     @BeforeEach
     void setUp() {
-        handlerMapping = new AnnotationHandlerMapping("samples.possible");
+        handlerMapping = new AnnotationHandlerMapping("samples");
         handlerMapping.initialize();
     }
