+++ mvc/src/test/java/samples/TestController.java
@@ -1,4 +1,4 @@
-package samples.possible;
+package samples;
 
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
